Update to file saving
Saving SHOULD work but it doesn't seem to be working because of some null reference.
Also I hit a roadblock because I'm not sure whether the player should be instantiated when starting a level or just moved to the wanted position and what to do with hud texts in case we instantiate the player.