Player head bobbing and changing FOV
All changes are in PlayerMove-script in a new method called CameraMovement.
FOV can match the player's speed, higher speeds equal higher field of view. However, this is currently disabled because it looks terrible.
Two applications for head bobbing. One is simple and works great, I made the other myself and it's better but has weird transitioning issues. This also needs quite a bit of tweaking.